Separate functions and definitions from usbdevice_fs into separate classes
